Use dynamic "version numbers" to bypass the browser cache of JavaScript and CSS.
When HTML requests JavaScript or CSS, the HTML will contain a version number (4 hex digits) for the requested file.
This version number is based on the last-modified date of the file on the server, so it should change each time
the file is updated. When the version number changes, the browser cache for that file becomes irrelevant and the
new version is fetched.